+++ packages/otel-collector/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @hyperdx/otel-collector
 
+## 2.23.0
+
 ## 2.22.1
 
 ### Patch Changes